Intigration_app/requirements.txt
Flask
openai
yt-dlp
whisper
moviepy
python-dotenv
pandas
scikit-learn
joblib
numpy
moviepy
youtube-transcript-api
PyPDF2
beautifulsoup4
requests
flask-cors